commit
Committing the changes

+++ src/app/modules/admin/example/currency/currency.component.ts
@@ -18,9 +18,11 @@ import { AddcurrencyComponent } from '../addcurrency/addcurrency.component';
 import { UpdatecurrencyComponent } from '../updatecurrency/updatecurrency.component';
 import { UpdatecurrencyService } from '../updatecurrency/updatecurrency.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { FuseConfirmationService } from '@fuse/services/confirmation';
 
 
 import { DataService } from '../../../../services/data.service';
+import { auto } from '@popperjs/core';
 
 
 export interface Currency {
@@ -30,7 +32,7 @@ export interface Currency {
   currencyName: string;
   buyRate: number;
   sellRate: number;
-  action: string;
+  // action: string;
 }
 
 @Component({
@@ -44,7 +46,7 @@ export interface Currency {
 
 
 export class CurrencyComponent implements OnInit {
-  displayedColumns: string[] = ['countryName', 'buyRate', 'sellRate', 'action'];
+  displayedColumns: string[] = ['countryName', 'buyRate', 'sellRate'];
   searchText: string = '';
   currentDate: Date = new Date();
   displayedData: Currency[] = [];
@@ -61,7 +63,8 @@ export class CurrencyComponent implements OnInit {
     private dialog:MatDialog,
     private updatecurrencyService: UpdatecurrencyService,
     private snackBar: MatSnackBar,
-    private dataService: DataService
+    private dataService: DataService,
+    private _fuseConfirmationService: FuseConfirmationService
   ) {}
 
   ngOnInit(): void {
@@ -81,7 +84,7 @@ export class CurrencyComponent implements OnInit {
 
 
   getService(){
-    this.currencyService.getCurrenciesByTenant(this.tenantId).subscribe((resp:any)=>{
+    this.currencyService.getData().subscribe((resp:any)=>{
       if(resp){
        this.currencies = resp; 
       this.displayedData = this.currencies;
@@ -106,14 +109,13 @@ export class CurrencyComponent implements OnInit {
     this.filteredData.data = this.displayedData;
   }
 
-
-  addCurrency()
-  {
-    this.dialog.open(AddcurrencyComponent,
-    {
-  disableClose: true,
-    })
-  }
+      dialogBoxSettings = {
+        height: auto,
+        width: '700px',
+        margin: '0 auto',
+        disableClose: true,
+        hasBackdrop: true
+      };
 
 
   openUpdateCurrencyDialog(countryId: string): void {
@@ -128,8 +130,38 @@ export class CurrencyComponent implements OnInit {
     });
   }
 
+
+
   deleteCurrency(countryId: string): void {
-    this.currencyService.deleteCurrencyByCountryId(countryId).subscribe(() => {
+  const confirmation = this._fuseConfirmationService.open({
+    title: 'Delete Currency',
+    message: 
+        'Are you sure you want to delete this currency?',
+    actions: {
+        confirm: {
+            label: 'Delete',
+        },
+        cancel: {
+          show: true,
+          label: 'Cancel',
+      },
+    },
+});
+
+// Subscribe to the confirmation dialog closed action
+confirmation.afterClosed().subscribe((result) => {
+
+    // If the confirm button pressed...
+
+    if (result === 'confirmed') {
+        // var CreatedBy = parseInt(localStorage.getItem("LoginId"))
+        // var data = {
+        //     ClientId: Id,
+        //     UpdatedBy: parseInt(localStorage.getItem("LoginId"))
+        // }
+
+        // Delete the currency
+     this.currencyService.deleteCurrencyByCountryId(countryId).subscribe(() => {
       console.log('Deleted Successfully.');
           // Show Snackbar Notification
           this.snackBar.open('Currency Deleted!', 'Close', {
@@ -142,7 +174,24 @@ export class CurrencyComponent implements OnInit {
     }, (error) => {
       console.log('Failed to delete');
     });
-  }
+    }
+});
+
+  // deleteCurrency(countryId: string): void {
+  //   this.currencyService.deleteCurrencyByCountryId(countryId).subscribe(() => {
+  //     console.log('Deleted Successfully.');
+  //         // Show Snackbar Notification
+  //         this.snackBar.open('Currency Deleted!', 'Close', {
+  //         duration: 3000, // Time in milliseconds
+  //         verticalPosition: 'top', // Position (top/bottom)
+  //         horizontalPosition: 'right', // Position (start/center/end/right/left)
+  //         panelClass: ['snackbar-success'] // Custom styling
+  //         });
+  //         this.getService();
+  //   }, (error) => {
+  //     console.log('Failed to delete');
+  //   });
+  // }
 
 
 
@@ -169,39 +218,4 @@ export class CurrencyComponent implements OnInit {
   // }
 
 }
-
-
-// let createOrEditSiteUserDialog;
-//         if (id === undefined || id <= 0) {
-//             createOrEditSiteUserDialog = this._dialog.open(CreateSiteuserComponent);
-//         } else {
-//             createOrEditSiteUserDialog = this._dialog.open(EditSiteuserComponent, {
-//                 data: id
-//             });
-//         }
-
-//         createOrEditSiteUserDialog.afterClosed().subscribe(result => {
-//             if (result) {
-//                // this.refresh();
-//                this.getgridFilter1()
-//             }
-//         });
-
-
-// const ELEMENT_DATA: Exchange[] = [
-//   { country: 'USD', code: 'US', currency_name: 'United States Dollar', buyRate: 58.3500, sellRate: 58.3500 },
-//   { country: 'GBP', code: 'GB', currency_name: 'British Pound Sterling', buyRate: 64.1400, sellRate: 66.0600 },
-//   { country: 'EUR', code: 'EU', currency_name: 'Euro', buyRate: 56.7400, sellRate: 58.1600 },
-//   { country: 'KSA', code: 'SA', currency_name: 'Saudi Riyal', buyRate: 23.3113, sellRate: 22.9301 },
-//   { country: 'UAE', code: 'AE', currency_name: 'United Arab Emirates Dirham', buyRate: 23.7988, sellRate: 23.4146 },
-//   { country: 'OM', code: 'OM', currency_name: 'Omani Rial', buyRate: 225.463, sellRate: 223.93500 },
-//   { country: 'AUD', code: 'AU', currency_name: 'Australian Dollar', buyRate: 37.1800, sellRate: 38.5100 },
-//   { country: 'BHD', code: 'BH', currency_name: 'Bahraini Dinar', buyRate: 130.8500, sellRate: 157.3400 },
-//   { country: 'CNY', code: 'CN', currency_name: 'Chinese Yuan', buyRate: 7.5926, sellRate: 8.3010 },
-//   { country: 'AED', code: 'AE', currency_name: 'United Arab Emirates Dirham', buyRate: 13.5000, sellRate: 16.0400 },
-//   { country: 'BND', code: 'BN', currency_name: 'Brunei Dollar', buyRate: 35.1700, sellRate: 41.1300 },
-//   { country: 'CHF', code: 'CH', currency_name: 'Swiss Franc', buyRate: 58.5400, sellRate: 60.6100 },
-//   { country: 'CAD', code: 'CA', currency_name: 'Canadian Dollar', buyRate: 41.7500, sellRate: 43.2100 },
-//   { country: 'HKD', code: 'HK', currency_name: 'Hong Kong Dollar', buyRate: 7.3570, sellRate: 7.5344 },
-//   { country: 'IDR', code: 'ID', currency_name: 'Indonesian Rupiah', buyRate: 0.0032, sellRate: 0.0039 },
-// ];
+}
